CI fix: Upgraded to artifact v4 (#144)
Artifact v3 has been removed, we updated to v4 and fixed the UB that was silently happening. Additionally we updated the runner images and the dependencies such as litmus, utfcpp, fmt, etc..

Lastly this PR exposed issues with CodeQL both in the false positives, some still visible in this PR, and the noisiness of Github Copilot autoreview that it had enabled. Resulting in a watershed of bad codegen. Removing copilot luckily removed those comments in the PR.

* Upgraded to artifact v4
* Resolved uniqueness of artifact name on windows
* Updated runner images
* updated external dependencies
* Updated clang format & resolved string literal issue
* removed codeql, more trouble than it is worth and weird terrible output with copilot