ECS add support for alternate mutability modes for components (#156)
This PR does some minor housekeeping and adds the following 2 features:
- Versioning support for components allowing you to tag, at compile
time, the version your component has (defaults to `0`). When
deserializing this version will be checked when the type information is
available.
- Alternate mutability modes for components

By default any component is available to modify from anywhere at any
time, this means that tracking changes such as through the
`on_condition` or by manual tracking in systems is expensive. The
`psl::ecs::component_mutability_behaviour_t::restricted` changes this by
disallowing write access to components which set this trait. These
components can only be modified through the
`psl::ecs::state_t::mutate_component` method which gives an instruction
(either directed to the specified fields), or through a complex object,
what parts of a component should mutate.

This mutation is automatically applied by the `state_t` instance when
`tick` is invoked, and systems which listen to `on_mutate<T>` (where `T`
is your given component type) will receive the modified components.

This allows for much cheaper reactivity to changed data, and narrows the
scope of _when_ data can be changed (but not by whom).

As an example usage, you could lock down material component changes
(such as an exposed colour parameter), and listen to the
`on_mutate<material_component>` to then only update the draw
instructions for the given instance of the changed material.