psl::ecs::on_mutate<> mutation information. (#157)
This improvement will allow you to query the `on_mutate<>` data to see
which fields have modified. Previously this parameter in the pack
contained the component data, you will need to query it separately now.

Additionally as this changes the components backing data mutations are
no longer applied during the serialization process.