Assembler inclusion (#160)
These sets of changes (together with modifications in the assembler)
made it possible for the assembler to be included with the engine for
the sake of the example data files generation.

Now when the examples are built the assembler is pulled in, and in the
case of the webgpu backend being required they will share their dep
setup.

Additionally this PR migrated some (most) libs to be enforced staticly
linked.

We also updated the dawn dependency to be in-sync with assembler (came
with an API upgrade for the webgpu's stub backend). We should find a
better solution for this down the road.