Instanced data rewrite (#175)
Resolves #173

Changes the backing of instanced data to be a psl::sparse_array (specifically an indices array). This allows us to hand out instance ID's which are disconnected from the memory locations allowing compaction etc..

Additionally this simplified several of the API's, and came with a performance boost (though lower peak as it's hard to beat 1:1 access to memory indices).

There were a few issues resolved in psl::sparse_array and psl::ecs::details::staged_sparse_array as well

* Instancing now using a sparse_array internally to manage to offsetting
* Added some missing utility functions needed for instancing specialization
* Utility to turn an indirect array into an array
* removed unneeded logging in ivk::buffer
* Added thread safety guard/checker to psl::sparse_array & psl::ecs::details::staged_sparse_array
* Fixed an incorrect behaviour when erasing from the dense storage for sparse_array
* Made staged_sparse_array::assign thread safe, and components_container::copy_from now uses that facility
* Fixed an issue caused by incorrectly calculating the chunk_index in sparse_array
* Added conditionality to the thread_safety_guard to allow debug-only variation
* Have a predefined size type for instancing defined at `core::gfx::instancing_size_type`
* Merged `static/dynamic _ add/remove` as they are functionaly the same now